requirements.txt
Django==3.1.7
psycopg2==2.8.4

# tests
pytest==6.2.1
pytest-sugar==0.9.4
pytest-django==4.1.0

# lint
black==20.8b1

django-ninja
google-auth-httplib2
google-auth-oauthlib
oauth2client==4.1.3
gspread==4.0.1
dateparser
deta
pandas
tinydb
pandas
gunicorn

honcho==1.0.1
huey[backends]==2.4.1
redis

